move like button to its own component to avoid rerendering the whole page
